server/requirements.txt
fastapi
uvicorn[standard]
dotenv
google-genai